V3.0 - Config.sh + APEX + ATP ECPU

+++ option/compute/compute_init.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
-# compute_java_bootstrap 
+# compute_init.sh 
 #
-# Script that is runned once during the setup of a 
-# - compute
-# - with Java
+# Init of a compute
+#
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+cd $SCRIPT_DIR
+
 if [[ -z "$TF_VAR_language" ]]; then
   echo "Missing env variables"
   exit
@@ -69,8 +71,11 @@ export -f install_java
 # -- App --------------------------------------------------------------------
 # Application Specific installation
 # Build all app* directories
+cd $HOME
+
 for APP_DIR in `ls -d app* | sort -g`; do
   if [ -f $APP_DIR/install.sh ]; then
+    echo "-- $APP_DIR: Install -----------------------------------------"
     chmod +x ${APP_DIR}/install.sh
     ${APP_DIR}/install.sh
   fi  
@@ -79,6 +84,7 @@ done
 # -- app/start.sh -----------------------------------------------------------
 for APP_DIR in `ls -d app* | sort -g`; do
   if [ -f $APP_DIR/start.sh ]; then
+    echo "-- $APP_DIR: Start -----------------------------------------"
     # Hardcode the connection to the DB in the start.sh
     if [ "$DB_URL" != "" ]; then
       sed -i "s!##JDBC_URL##!$JDBC_URL!" $APP_DIR/start.sh 
@@ -111,24 +117,30 @@ EOT
   fi
 done  
 
+# -- Helper --------------------------------------------------------------------
+cd $SCRIPT_DIR
+mv helper.sh $HOME
+
 # -- UI --------------------------------------------------------------------
 # Install NGINX
 sudo dnf install nginx -y > /tmp/dnf_nginx.log
 
 # Default: location /app/ { proxy_pass http://localhost:8080 }
-sudo cp nginx_app.locations /etc/nginx/conf.d/.
-if grep -q nginx_app /etc/nginx/nginx.conf; then
-  echo "Include nginx_app.locations is already there"
-else
+if [ -f nginx_app.locations ]; then
+  sudo cp nginx_app.locations /etc/nginx/conf.d/.
+  if grep -q nginx_app /etc/nginx/nginx.conf; then
+    echo "Include nginx_app.locations is already there"
+  else
     echo "Adding nginx_app.locations"
     sudo awk -i inplace '/404.html/ && !x {print "        include conf.d/nginx_app.locations;"; x=1} 1' /etc/nginx/nginx.conf
+  fi
 fi
 
 # TLS
 if [ -f nginx_tls.conf ]; then
-    echo "Adding nginx_tls.conf"
-    sudo cp nginx_tls.conf /etc/nginx/conf.d/.
-    sudo awk -i inplace '/# HTTPS server/ && !x {print "        include conf.d/nginx_tls.conf;"; x=1} 1' /etc/nginx/nginx.conf
+  echo "Adding nginx_tls.conf"
+  sudo cp nginx_tls.conf /etc/nginx/conf.d/.
+  sudo awk -i inplace '/# HTTPS server/ && !x {print "        include conf.d/nginx_tls.conf;"; x=1} 1' /etc/nginx/nginx.conf
 fi
 
 # SE Linux (for proxy_pass)
@@ -138,6 +150,7 @@ sudo setsebool -P httpd_can_network_connect 1
 sudo systemctl enable nginx
 sudo systemctl restart nginx
 
+cd $HOME
 if [ -d ui ]; then
   # Copy the index file after the installation of nginx
   sudo cp -r ui/* /usr/share/nginx/html/
